Optimize network requests on map view page
Fixes #93
